requirements.txt
argh==0.26.2
paho-mqtt==1.3.1
pathtools==0.1.2
pymongo==3.5.1
PyYAML==3.12
watchdog==0.8.3
